\name{mousemove}
\docType{data}
\alias{mousemove}
    
\title{Mouse movement trajectory}

\description{
  
  A single trial for a mouse-movement trajectory (x, y, time),
  and the type of counterbalancing used ("Left") in this specific 
  case. 
  
}

\usage{mousemov}
\format{A dataframe with 4 columns (x, y, time, counterbl), and 
  61 rows.}

\references{

  Coco, M.I., and Duran, D.N. (under review). Action-dynamics 
  as revision costs: the interaction of plausibility and congruency. 

}

\keyword{datasets}
